+++ packages/terra-form-select/tests/wdio/select-spec.js
@@ -2044,4 +2044,41 @@ Terra.describeViewports('Select', ['tiny'], () => {
 
     Terra.it.validatesElement();
   });
+
+  describe('With-Iframe should not cause page jump when', () => {
+    before(() => browser.url('/#/raw/tests/terra-form-select/form-select/select-inside-iframe'));
+
+    it('open the dropdown by clicking the toggle', () => {
+      /* global $ */
+      const frame = $('#frmSelectFrame');
+      frame.waitForVisible();
+      browser.frame(frame.value);
+      browser.click('#frmSelectSingle');
+      browser.moveToObject('#frmSelectSingle');
+    });
+
+    Terra.it.validatesElement('opened form-select', { selector: '#root' });
+
+    it('select the option which is hidden from visible frame of form-select', () => {
+      browser.moveToObject('#terra-select-option-17');
+      browser.click('#terra-select-option-17');
+      browser.click('#frmSelectSingle');
+    });
+
+    Terra.it.validatesElement('opened form-select with selected-option', { selector: '#root' });
+
+    it('navigates to next option of form-select on down arrow key oress', () => {
+      browser.keys('ArrowDown');
+      browser.keys('ArrowDown');
+    });
+
+    Terra.it.validatesElement('navigated to next option using key-board navigation', { selector: '#root' });
+
+    it('navigates to next option of form-select on down arrow key oress', () => {
+      browser.keys('ArrowDown');
+      browser.keys('ArrowUp');
+    });
+
+    Terra.it.validatesElement('navigated to previous option using key-board navigation', { selector: '#root' });
+  });
 });