+++ packages/terra-table/src/index.js
@@ -1,0 +1,7 @@
+import Table from './Table';
+import Utils from './TableUtils';
+
+export default Table;
+export {
+  Utils,
+};