+++ packages/terra-paginator/CHANGELOG.md
@@ -3,6 +3,11 @@ ChangeLog
 
 Unreleased
 ----------
+### Added
+* Private PaginatorButton component added to separate button focus and active style management
+
+### Changed
+* `focused` and `active` pseudo selectors migrated to data attributes
 
 2.37.0 - (January 28, 2020)
 ------------------